Handle package renames in qualified class name references
The 'package' rename is now for the referenced class, instead of the file,
and includes package references anywhere within the file, not just the
top-level.

For example, in CraftInventoryPlayer, net.minecraft.server is now renamed
to net.minecraft.entity.player (first use, referencing InventoryPlayer).

Fixes 90 errors in CB (669 -> 579), mostly in OBC.